Moved funcs to utils and log that are needed in front-end #36
