+++ odm360/utils.py
@@ -4,6 +4,7 @@ import logging
 logger = logging.getLogger(__name__)
 import socket
 import nmap
+from configparser import ConfigParser
 
 
 def find_serial(wildcard='', logger=logger):
@@ -48,3 +49,12 @@ def get_lan_devices(ip_subnet):
     nm = nmap.PortScanner()  # instantiate nmap.PortScanner object
     nm.scan(hosts=ip_subnet, arguments='-sP')
     return [(x, nm[x]['status']['state']) for x in nm.all_hosts()]
+
+def parse_config(settings_path):
+    """
+        Parse the config file with interpolation=None or it will break run_cast.sh
+    """
+    config = ConfigParser(interpolation=None)
+    config.read(settings_path)
+    return config
+
